// src/payload-types.ts
export interface Media {
  id: number;
  alt?: string | null;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
}
// ...
export interface Timeline {
  id: number;
  title: string;
  date: string;
  description?: string | null;
  media?:
    | (
        | {
            url: string;
            id?: string | null;
            blockName?: string | null;
            blockType: 'youtubeLink';
          }
        | {
            image: number | Media;
            id?: string | null;
            blockName?: string | null;
            blockType: 'imageMedia';
          }
      )[]
    | null;
  link?: string | null;
  label?: string | null;
  updatedAt: string;
  createdAt: string;
}